requirements.txt
transformers==4.25.1
datasets==2.0.0
torch --extra-index-url https://download.pytorch.org/whl/cu116
evaluate==0.3.0
pyvi==0.1.1
redis==3.4.1
msgpack==0.5.6
coloredlogs==10.0
gunicorn==20.0.4
flask_cors==3.0.10
flask==2.0.2
Cython==0.29.10
pyaml==20.4.0
